refactor: 💡 Move stream length input into its own component
